refactor: move scroll margin calculation to separate hook
